+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.1.1]
+
+### Added
+ - Added `isSameAs()` method to DateRange
+
+
+
 ## [2.1.0]
 
 ### Added